Rename the auxtel*checkout scripts to auxtel_daytime* to better organize them.
Update init module.
