+++ python/lsst/ts/IntegrationTests/auxtel_daytime_latiss.py
