+++ python/lsst/ts/IntegrationTests/auxtel_daytime_slew_and_take_image.py
