+++ python/lsst/ts/IntegrationTests/auxtel_daytime_telescope_and_dome.py
